requirements.txt
langchain==0.1.4
streamlit==1.32.2
streamlit-authenticator==0.3.2
python-dotenv==1.0.1
langchain-openai==0.1.3